third_party/WebKit/LayoutTests/inspector/console/console-uncaught-promise-in-worker-expected.txt

CONSOLE ERROR: line 3: Uncaught (in promise) Error: err1
CONSOLE ERROR: line 21: Uncaught (in promise) Error: err2
Tests that uncaught promise rejections happenned in workers are logged into console.

worker-with-unhandled-promises.js:3 Uncaught (in promise) Error: err1 {stack: (...), message: "err1"}
    message: "err1"
    stack: (...)
    get stack: function () { [native code] }
    set stack: function () { [native code] }
    __proto__: d
worker-with-unhandled-promises.js:21 Uncaught (in promise) Error: err2 {stack: (...), message: "err2"}
    message: "err2"
    stack: (...)
    get stack: function () { [native code] }
    set stack: function () { [native code] }
    __proto__: d
